requirements.txt
allure-pytest
requests
pytest
pytest-xdist
pytest-order
python-dotenv
